+++ lib/rspec/rails/example/rails_example_group.rb
@@ -16,6 +16,7 @@ module RSpec
       include RSpec::Rails::MinitestAssertionAdapter
       include RSpec::Rails::FixtureSupport
       include RSpec::Rails::TaggedLoggingAdapter
+      include ActiveJob::TestHelper if RSpec::Rails::FeatureCheck.has_active_job?
       include ActiveSupport::CurrentAttributes::TestHelper
       include ActiveSupport::ExecutionContext::TestHelper
     end